Add linkage with SG5 TM
I wanted to add how executors interact with SG5 TM in the proposal but ran out of time. We at SG5 is starting to study his interaction, and I wanted to put a placeholder here (actually in the last proposal but I ran out of time)  in case anyone wonders why interaction with a concurrency mechanism like TM is not stated.